Locations reorg
