+++ client/web/src/components/MapPage/Locations/LocationDetail/index.js
