+++ client/web/src/components/MapPage/Locations/LocationsList/index.js
