+++ client/web/src/components/MapPage/Locations/LocationsSummary.js
